requirements.txt
beautifulsoup4
matplotlib
numpy
pandas
pypfopt
python-dateutil
pyyaml
selenium
sqlalchemy
talib
